+++ api/adm_checkForAdmin.php
@@ -32,7 +32,7 @@
         } else {
             echo "false";
         }
-    }
+    } 
     // show error
     catch (Exception $e){